Fix Kotlin Multiplatform support
