+++ kopy-args/build.gradle.kts
@@ -16,6 +16,39 @@ hubdle {
         }
     }
     kotlin {
-        jvm() //
+        multiplatform {
+            common()
+            androidNative()
+            apple {
+                ios {
+                    iosArm64()
+                    iosSimulatorArm64()
+                    iosX64()
+                }
+                macos {
+                    macosArm64()
+                    macosX64()
+                }
+                tvos {
+                    tvosArm64()
+                    tvosSimulatorArm64()
+                    tvosX64()
+                }
+            }
+            jvm()
+            js {
+                browser()
+                nodejs()
+            }
+            linux {
+                linuxArm64()
+                linuxX64()
+            }
+            mingw {
+                mingwX64()
+            }
+            native()
+            wasmJs()
+        }
     }
 }
