+++ kopy-args/common/main/kotlin/com/javiersc/kotlin/kopy/args/KopyArgument.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.kopy.args
+
+public expect interface KopyArgument
