+++ kopy-args/js/main/kotlin/com/javiersc/kotlin/kopy/args/KopyArgument.js.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.kopy.args
+
+public actual interface KopyArgument
