+++ kopy-args/wasmJs/main/kotlin/com/javiersc/kotlin/kopy/args/KopyArgument.wasmJs.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.kopy.args
+
+public actual interface KopyArgument
